[7330] Code warnings and style cleanups. Some bugs fixes. Author: VladimirMangos
    1) comparison singed and unsigned values
    2) redundent includes
    3) wrong constructor :-part field initilization
    4) unused not-/*name*/-guarded args in template/virtual functions that not required like args.
    5) explicitly list not implemented achievement types.

    Also bugs fixed:
    1) Drop wrong phase mask 0 check in WorldObject::InSamePhase.
    2) ArenaTeamMember::ModifyPersonalRating incorrect work with move points in negative with infinity values in result.
    3) ArenaTeam::SaveToDB code send uint64 value to string with arg format %u.

--HG--
branch : trunk

+++ src/game/ArenaTeam.h
@@ -98,7 +98,7 @@ struct ArenaTeamMember
 
     void ModifyPersonalRating(Player* plr, int32 mod, uint32 slot)
     {
-        if (personal_rating + mod < 0)
+        if (int32(personal_rating) + mod < 0)
             personal_rating = 0;
         else
             personal_rating += mod;
